refactor: Move scheduler type classes to an own package
