+++ core/src/main/java/de/pxav/kelp/core/scheduler/type/SchedulerFactory.java
@@ -1,7 +1,8 @@
-package de.pxav.kelp.core.scheduler;
+package de.pxav.kelp.core.scheduler.type;
 
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
+import de.pxav.kelp.core.scheduler.KelpSchedulerRepository;
 
 /**
  * A class description goes here.
